+++ NVActivityIndicatorView/NVActivityIndicatorViewable.swift
@@ -11,65 +11,48 @@ import UIKit
 /**
  *  UIViewController conforms this protocol to be able to display NVActivityIndicatorView as UI blocker.
  *
- *  This will add functionalities to UIViewController to display and remove UI blocker.
+ *  This extends abilities of UIViewController to display and remove UI blocker.
  */
 public protocol NVActivityIndicatorViewable { }
 
 public extension NVActivityIndicatorViewable where Self: UIViewController {
-
-    private var activityRestorationIdentifier: String {
-        return "NVActivityIndicatorViewContainer"
-    }
-
+    
     /**
-     Create a activity indicator view with specified frame, type, color and padding and start animation.
+     Display UI blocker.
      
-     - parameter frame: view's frame.
-     - parameter type: animation type, value of NVActivityIndicatorType enum. Default type is BallSpinFadeLoader.
-     - parameter color: color of activity indicator view. Default color is white.
-     - parameter padding: view's padding. Default padding is 0.
+     Appropriate NVActivityIndicatorView.DEFAULT_* values are used for omitted params.
+     
+     - parameter size:                 size of activity indicator view.
+     - parameter message:              message displayed under activity indicator view.
+     - parameter type:                 animation type.
+     - parameter color:                color of activity indicator view.
+     - parameter padding:              padding of activity indicator view.
+     - parameter displayTimeThreshold: display time threshold to actually display UI blocker.
+     - parameter minimumDisplayTime:   minimum display time of UI blocker.
      */
-    public func startActivityAnimating(message: String? = nil, type: NVActivityIndicatorType? = nil, color: UIColor? = nil, padding: CGFloat? = nil) {
-        let activityContainer: UIView = UIView(frame: view.bounds)
-        activityContainer.backgroundColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.5)
-        activityContainer.restorationIdentifier = activityRestorationIdentifier
+    public func startAnimating(
+        _ size: CGSize? = nil,
+        message: String? = nil,
+        type: NVActivityIndicatorType? = nil,
+        color: UIColor? = nil,
+        padding: CGFloat? = nil,
+        displayTimeThreshold: Int? = nil,
+        minimumDisplayTime: Int? = nil) {
+        let activityData = ActivityData(size: size,
+                                        message: message,
+                                        type: type,
+                                        color: color,
+                                        padding: padding,
+                                        displayTimeThreshold: displayTimeThreshold,
+                                        minimumDisplayTime: minimumDisplayTime)
         
-        let width = activityContainer.frame.size.width / 3
-        let height = width
-        
-        let activityIndicatorView = NVActivityIndicatorView(
-            frame: CGRectMake(0, 0, width, height),
-            type: type,
-            color: color,
-            padding: padding)
-        activityIndicatorView.center = activityContainer.center
-        activityIndicatorView.hidesWhenStopped = true
-        activityIndicatorView.startAnimation()
-        activityContainer.addSubview(activityIndicatorView)
-        
-        if let message = message where !message.isEmpty {
-            let label = UILabel(frame: CGRectMake(0, 0, width, 30))
-            label.center = CGPointMake(
-                activityIndicatorView.center.x,
-                activityIndicatorView.center.y + height)
-            label.textAlignment = .Center
-            label.text = message
-            label.font = UIFont.boldSystemFontOfSize(20)
-            label.textColor = activityIndicatorView.color
-            activityContainer.addSubview(label)
-        }
-        
-        view.addSubview(activityContainer)
-    }
-
-    /**
-     Stop animation and remove from view hierarchy.
-     */
-    public func stopActivityAnimating() {
-        for item in view.subviews
-            where item.restorationIdentifier == activityRestorationIdentifier {
-                item.removeFromSuperview()
-        }
+        NVActivityIndicatorPresenter.sharedInstance.startAnimating(activityData)
     }
     
+    /**
+     Remove UI blocker.
+     */
+    public func stopAnimating() {
+        NVActivityIndicatorPresenter.sharedInstance.stopAnimating()
+    }
 }